requirements.txt
# Runtime dependencies
pydantic>=2.7
pydantic-settings>=2.3
typer>=0.12
rich>=13.7
python-dotenv>=1.0
PyYAML>=6.0
tenacity>=8.3

# Optional LLM / agent orchestration dependencies
openai>=1.40
langgraph>=0.2
langchain-core>=0.2
langsmith>=0.1
tavily-python>=0.3.0

# Dev / test dependencies
pytest>=8.2
pytest-cov>=5.0
ruff>=0.5
mypy>=1.10
pre-commit>=3.7
